WIP 4: Interceptor rework: break apart intercept module
